add sync_frame but not test

+++ app/src/main/java/yong/tank/tool/StaticVariable.java
@@ -24,7 +24,7 @@ public class StaticVariable {
     //设置游戏的关键帧增长率
     public static int KEY_FRAME_COUNT = 3;
     //设置关键帧的帧为： -----0-3-6-9-12-15-18-----
-    public static int KEY_FRAME = 0;
+    public static long KEY_FRAME = 0;
     //设置 游戏的最大跳过帧数为5：
     public static int  MAX_FRAMESKIP = 5;
     public static int SKIP_TICKS = 1000 / LOGICAL_FRAME;
